move wrappers to dedicated wrappers folder, mock withSession wrapper
