+++ src/app/api/auth/login/google/route.ts
@@ -1,8 +1,8 @@
 import 'server-only'
-import { createSessionCookie } from '@/app/api/session'
+import { createSessionCookie } from '@/app/api/wrappers/session'
 import { OAuth2Client } from 'google-auth-library'
 import { NextResponse, NextRequest } from 'next/server'
-import { withCSRFProtection } from '@/app/api/csrf'
+import { withCSRFProtection } from '@/app/api/wrappers/csrf'
 import sanitizeUserData from '@/app/api/utils/sanitizeUserData'
 import getResponseError from '@/app/api/utils/getResponseError'
 import getUserData from '../getUserData'
